refactor(dynamic-agents): import get_runtime_cache from runtime_cache directly
